pyproject.toml
[project]
name = "dteam-django-practical-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Kindrat22",email = "volodimirkindrat22@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "weasyprint (>=65.1,<66.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "celery",
    "redis",
    "django-celery-results"
]

packages = [
    { include = "main" },
    { include = "audit" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
